Code formatting

+++ script/deployment/DeploymentScript.sol
@@ -3,12 +3,10 @@ pragma solidity ^0.8.0;
 
 import 'forge-std/Script.sol';
 
-
 /**
  * Provides additional logic and helper methods for deployment scripts.
  */
 contract DeploymentScript is Script {
-
     /**
      * Defines the JSON structure used to store contract deployment information.
      *
@@ -88,10 +86,7 @@ contract DeploymentScript is Script {
 
         // Register our new deployment data
         if (!exists) {
-            newDeployments[newDeployments.length - 1] = Deployment({
-                key: key,
-                deploymentAddress: deploymentAddress
-            });
+            newDeployments[newDeployments.length - 1] = Deployment({key: key, deploymentAddress: deploymentAddress});
         } else {
             newDeployments = deployments;
         }
@@ -122,7 +117,7 @@ contract DeploymentScript is Script {
      */
     modifier deployer() {
         // Load our seed phrase from a protected file
-        uint256 privateKey = _stringToUint(vm.readFile('.privatekey'));
+        uint privateKey = _stringToUint(vm.readFile('.privatekey'));
 
         // Using the passed in the script call, has all subsequent calls (at this call
         // depth only) create transactions that can later be signed and sent onchain.
@@ -142,30 +137,29 @@ contract DeploymentScript is Script {
      *
      * @return string The address in string format
      */
-    function _addressToString(address _address) internal pure returns(string memory) {
-       bytes32 _bytes = bytes32(uint256(uint160(_address)));
-       bytes memory HEX = "0123456789abcdef";
-       bytes memory _string = new bytes(42);
-       _string[0] = '0';
-       _string[1] = 'x';
-       for(uint i = 0; i < 20; i++) {
-           _string[2+i*2] = HEX[uint8(_bytes[i + 12] >> 4)];
-           _string[3+i*2] = HEX[uint8(_bytes[i + 12] & 0x0f)];
-       }
+    function _addressToString(address _address) internal pure returns (string memory) {
+        bytes32 _bytes = bytes32(uint(uint160(_address)));
+        bytes memory HEX = '0123456789abcdef';
+        bytes memory _string = new bytes(42);
+        _string[0] = '0';
+        _string[1] = 'x';
+        for (uint i = 0; i < 20; i++) {
+            _string[2 + i * 2] = HEX[uint8(_bytes[i + 12] >> 4)];
+            _string[3 + i * 2] = HEX[uint8(_bytes[i + 12] & 0x0f)];
+        }
 
-       return string(_string);
+        return string(_string);
     }
 
     function _stringToUint(string memory s) internal pure returns (uint) {
         bytes memory b = bytes(s);
         uint result = 0;
-        for (uint256 i = 0; i < b.length; i++) {
-            uint256 c = uint256(uint8(b[i]));
+        for (uint i = 0; i < b.length; i++) {
+            uint c = uint(uint8(b[i]));
             if (c >= 48 && c <= 57) {
                 result = result * 10 + (c - 48);
             }
         }
         return result;
     }
-
 }
