+++ test/mocks/erc/ERC20Mock.sol
@@ -6,11 +6,11 @@ import {ERC20} from '@openzeppelin/contracts/token/ERC20/ERC20.sol';
 contract ERC20Mock is ERC20 {
     constructor() ERC20('ERC20Mock', 'E20M') {}
 
-    function mint(address account, uint256 amount) external {
+    function mint(address account, uint amount) external {
         _mint(account, amount);
     }
 
-    function burn(address account, uint256 amount) external {
+    function burn(address account, uint amount) external {
         _burn(account, amount);
     }
 }
